fixes tool voters (#6773)
