+++ main/core/Security/Voter/Tool/OrderedToolVoter.php
@@ -11,16 +11,9 @@
 
 namespace Claroline\CoreBundle\Security\Voter\Tool;
 
-use Claroline\AppBundle\Persistence\ObjectManager;
 use Claroline\CoreBundle\Entity\Tool\OrderedTool;
-use Claroline\CoreBundle\Entity\Tool\ToolRights;
-use Claroline\CoreBundle\Manager\Tool\ToolMaskDecoderManager;
-use Claroline\CoreBundle\Manager\Workspace\WorkspaceManager;
-use Claroline\CoreBundle\Repository\Tool\ToolRightsRepository;
 use Claroline\CoreBundle\Security\Voter\AbstractVoter;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
-use Symfony\Component\Security\Core\Authorization\Voter\VoterInterface;
-use Symfony\Component\Security\Core\Role\Role;
 
 /**
  * Checks if the current token can access a tool configured in a Workspace or Desktop
@@ -28,33 +21,6 @@ use Symfony\Component\Security\Core\Role\Role;
  */
 class OrderedToolVoter extends AbstractVoter
 {
-    /** @var ToolMaskDecoderManager */
-    private $maskManager;
-
-    /** @var WorkspaceManager */
-    private $workspaceManager;
-
-    /** @var ToolRightsRepository */
-    private $rightsRepository;
-
-    /**
-     * OrderedToolVoter constructor.
-     *
-     * @param ObjectManager          $om
-     * @param ToolMaskDecoderManager $maskManager
-     * @param WorkspaceManager       $workspaceManager
-     */
-    public function __construct(
-        ObjectManager $om,
-        ToolMaskDecoderManager $maskManager,
-        WorkspaceManager $workspaceManager
-    ) {
-        $this->maskManager = $maskManager;
-        $this->workspaceManager = $workspaceManager;
-
-        $this->rightsRepository = $om->getRepository(ToolRights::class);
-    }
-
     /**
      * @param TokenInterface $token
      * @param OrderedTool    $object
@@ -65,28 +31,13 @@ class OrderedToolVoter extends AbstractVoter
      */
     public function checkPermission(TokenInterface $token, $object, array $attributes, array $options)
     {
-        if ($this->isAdmin($token)) {
-            return VoterInterface::ACCESS_GRANTED;
+        if (!empty($object->getWorkspace())) {
+            // let the workspace voter decide
+            return $this->isGranted([$object->getTool()->getName(), $attributes[0]], $object->getWorkspace());
         }
 
-        if (!empty($object->getWorkspace()) && $this->workspaceManager->isManager($object->getWorkspace(), $token)) {
-            return VoterInterface::ACCESS_GRANTED;
-        }
-
-        $decoder = $this->maskManager->getMaskDecoderByToolAndName($object->getTool(), $attributes[0]);
-        if ($decoder) {
-            $mask = $this->rightsRepository->findMaximumRights(array_map(function (Role $role) {
-                return $role->getRole();
-            }, $token->getRoles()), $object);
-
-            if ($mask & $decoder->getValue()) {
-                return VoterInterface::ACCESS_GRANTED;
-            }
-
-            return VoterInterface::ACCESS_DENIED;
-        }
-
-        return VoterInterface::ACCESS_ABSTAIN;
+        // let the base tool voter decide
+        return $this->isGranted($attributes[0], $object->getTool());
     }
 
     public function getClass()
